refactor: move to more generic namespace
