+++ Components/Navigation/IDockable.cs
@@ -1,6 +1,3 @@
-using Godot;
-using Spacelancer.Scenes.Player;
-
 namespace Spacelancer.Components.Navigation;
 
 public interface IDockable : INavigable
