+++ Components/Navigation/LaneNavigation.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using Godot;
 using Serilog;
-using Spacelancer.Scenes.Player;
-
 
 namespace Spacelancer.Components.Navigation;
