+++ Components/Navigation/PlayerNavigation.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Godot;
 using Serilog;
+using Spacelancer.Scenes.SpaceShips;
 
 namespace Spacelancer.Components.Navigation;
 
@@ -10,9 +11,9 @@ public class PlayerNavigation : INavigationSoftware
     public NavigationSoftwareType Type => NavigationSoftwareType.Manual;
 
     private const string _name = "PlayerNavigation";
-    private readonly Scenes.Player.Player _player;
+    private readonly Player _player;
 
-    public PlayerNavigation(Scenes.Player.Player player)
+    public PlayerNavigation(Player player)
     {
         _player = player;
     }
