+++ Controllers/Global.cs
@@ -1,4 +1,5 @@
 using Godot;
+using Spacelancer.Scenes.SpaceShips;
 
 namespace Spacelancer.Controllers;
 
@@ -14,7 +15,7 @@ public partial class Global : Node
     public static EconomyController Economy { get; } = new();
     public static UniverseController Universe { get; } = new();
     public static UiController UserInterface { get; private set; }
-    public static Scenes.Player.Player Player;
+    public static Player Player;
     
     internal Logger Logger;
     internal SettingsController SettingsController;
